Choose correct quarter for events.html with no future events
Previously, if there were no future events during a compile, the code
that generates events.html as a copy of the "current" events page
would crash, as it used the upcoming future event as the way of
deciding which events page is "current". This changes the code to
gracefully handle this situation, defaulting to showing
the (chronologically) greatest events page if there are no future
events. This is probably the correct behavior, because when there are
no future events, it is probably because a quarter has just finished
and the next has not yet started; the events page from the previous
quarter will be the chronologically greatest, as well as the most
recent/relevant.