Use events/past instead of future tags to decide event status
Previously, an event was shown on the front page, and used to
calculate the "current" quarter of events, if and only if it contained
a tag named "future" in its yaml metadata block. This changes site.hs
to instead include an event if and only if its identifier is not
included in events/past, a record of which events have been completed.