Fix datepicker to create pandoc-compatible files
Previously, the date-picker markdown file generator generated files that
contained a UTF-8 BOM.  Unfortunately, Pandoc does not recognize the
yaml metadata block in files that contain a UTF-8 BOM, and so was unable
to process markdown files generated by the automated generator.  This
modifies the generator to ensure that files are generated without a
UTF-8 BOM, and also ensures that another extraneous newline is removed.